+++ Samples~/AR Starter Assets/ARDemoSceneAssets/Scripts/ARSampleMenuManager.cs
@@ -1,5 +1,4 @@
 using UnityEngine.EventSystems;
-using UnityEngine.InputSystem;
 using UnityEngine.UI;
 using UnityEngine.XR.Interaction.Toolkit.Samples.StarterAssets;
 
@@ -116,13 +115,10 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
             set => m_InteractionGroup = value;
         }
 
-        bool m_IsPointerOverUI;
         bool m_ShowObjectMenu;
 
         void OnEnable()
         {
-            m_ScreenSpaceController.dragCurrentPositionAction.action.started += HideTapOutsideUI;
-            m_ScreenSpaceController.tapStartPositionAction.action.started += HideTapOutsideUI;
             m_CreateButton.onClick.AddListener(ShowMenu);
             m_CancelButton.onClick.AddListener(HideMenu);
             m_DeleteButton.onClick.AddListener(DeleteFocusedObject);
@@ -131,8 +127,6 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
         void OnDisable()
         {
             m_ShowObjectMenu = false;
-            m_ScreenSpaceController.dragCurrentPositionAction.action.started -= HideTapOutsideUI;
-            m_ScreenSpaceController.tapStartPositionAction.action.started -= HideTapOutsideUI;
             m_CreateButton.onClick.RemoveListener(ShowMenu);
             m_CancelButton.onClick.RemoveListener(HideMenu);
             m_DeleteButton.onClick.RemoveListener(DeleteFocusedObject);
@@ -149,17 +143,21 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
             {
                 m_CreateButton.gameObject.SetActive(false);
                 m_DeleteButton.gameObject.SetActive(false);
-                m_IsPointerOverUI = EventSystem.current != null && EventSystem.current.IsPointerOverGameObject(-1);
+                var isPointerOverUI = EventSystem.current != null && EventSystem.current.IsPointerOverGameObject(-1);
+                if (!isPointerOverUI && m_ScreenSpaceController.tapStartPositionAction.action.WasPerformedThisFrame())
+                {
+                    HideMenu();
+                }
             }
-            else
+            else if (m_InteractionGroup is not null)
             {
-                m_IsPointerOverUI = false;
-                if (m_InteractionGroup?.focusInteractable != null && (!m_DeleteButton.isActiveAndEnabled || m_CreateButton.isActiveAndEnabled))
+                var currentFocusedObject = m_InteractionGroup.focusInteractable;
+                if (currentFocusedObject != null && (!m_DeleteButton.isActiveAndEnabled || m_CreateButton.isActiveAndEnabled))
                 {
                     m_CreateButton.gameObject.SetActive(false);
                     m_DeleteButton.gameObject.SetActive(true);
                 }
-                else if (m_InteractionGroup?.focusInteractable == null && (!m_CreateButton.isActiveAndEnabled || m_DeleteButton.isActiveAndEnabled))
+                else if (currentFocusedObject == null && (!m_CreateButton.isActiveAndEnabled || m_DeleteButton.isActiveAndEnabled))
                 {
                     m_CreateButton.gameObject.SetActive(true);
                     m_DeleteButton.gameObject.SetActive(false);
@@ -171,17 +169,17 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
         {
             if (m_ObjectSpawner == null)
             {
-                Debug.LogWarning("Object Spawner not configured correctly: no ObjectSpawner set.");
+                Debug.LogWarning("Menu Manager not configured correctly: no Object Spawner set.", this);
             }
             else
             {
-                if (m_ObjectSpawner.objectPrefabs.Count > objectIndex)
+                if (objectIndex < m_ObjectSpawner.objectPrefabs.Count)
                 {
                     m_ObjectSpawner.spawnOptionIndex = objectIndex;
                 }
                 else
                 {
-                    Debug.LogWarning("Object Spawner not configured correctly: object index larger than number of Object Prefabs.");
+                    Debug.LogWarning("Object Spawner not configured correctly: object index larger than number of Object Prefabs.", this);
                 }
             }
 
@@ -207,16 +205,11 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
             m_ShowObjectMenu = false;
         }
 
-        void HideTapOutsideUI(InputAction.CallbackContext context)
-        {
-            if (!m_IsPointerOverUI)
-            {
-                HideMenu();
-            }
-        }
-
         void DeleteFocusedObject()
         {
+            if (m_InteractionGroup == null)
+                return;
+
             var currentFocusedObject = m_InteractionGroup.focusInteractable;
             if (currentFocusedObject != null)
             {