+++ Samples~/Hands Interaction Demo/Scripts/MetaSystemGestureDetector.cs
@@ -130,7 +130,7 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.Hands
 
         readonly BindableEnum<SystemGestureState> m_SystemGestureState = new BindableEnum<SystemGestureState>(checkEquality: false);
 
-#if XR_HANDS_1_1_OR_NEWER
+#if XR_HANDS_1_1_OR_NEWER && (ENABLE_VR || UNITY_GAMECORE)
         [NonSerialized] // NonSerialized is required to avoid an "Unsupported enum base type" error about the Flags enum being ulong
         MetaAimFlags m_AimFlags;
 #endif
@@ -145,10 +145,12 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.Hands
             BindAimFlags();
 
 #if XR_HANDS_1_1_OR_NEWER
+#if ENABLE_VR || UNITY_GAMECORE
             var action = m_AimFlagsAction.action;
             if (action != null)
                 // Force invoking the events upon initialization to simplify making sure the callback's desired results are synced
                 UpdateAimFlags((MetaAimFlags)action.ReadValue<int>(), true);
+#endif
 #else
             Debug.LogWarning("Script requires XR Hands (com.unity.xr.hands) package to monitor Meta Aim Flags. Install using Window > Package Manager or click Fix on the related issue in Edit > Project Settings > XR Plug-in Management > Project Validation.", this);
             SetGestureState(SystemGestureState.Ended, true);
@@ -212,7 +214,7 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.Hands
             }
         }
 
-#if XR_HANDS_1_1_OR_NEWER
+#if XR_HANDS_1_1_OR_NEWER && (ENABLE_VR || UNITY_GAMECORE)
         void UpdateAimFlags(MetaAimFlags value, bool forceInvoke = false)
         {
             var hadMenuPressed = (m_AimFlags & MetaAimFlags.MenuPressed) != 0;
@@ -253,7 +255,7 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.Hands
 
         void OnAimFlagsActionPerformedOrCanceled(InputAction.CallbackContext context)
         {
-#if XR_HANDS_1_1_OR_NEWER
+#if XR_HANDS_1_1_OR_NEWER && (ENABLE_VR || UNITY_GAMECORE)
             UpdateAimFlags((MetaAimFlags)context.ReadValue<int>());
 #endif
         }